+++ catalog/src/main/java/it/eng/catalog/event/CatalogEventListener.java
@@ -2,7 +2,6 @@ package it.eng.catalog.event;
 
 import it.eng.catalog.model.Catalog;
 import it.eng.catalog.service.CatalogService;
-import it.eng.tools.event.contractnegotiation.ContractNegotationOfferRequestEvent;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -21,12 +20,6 @@ public class CatalogEventListener {
 	@EventListener
 	public void handleContextStart(Catalog catalog) {
 		log.info("Handling context started event. " + catalog.getId());
-	}
-	
-	@EventListener
-	public void handleContractNegotationOfferRequestEvent(ContractNegotationOfferRequestEvent offerRequest) {
-		log.info("Received event - ContractNegotationOfferRequestEvent");
-		catalogService.validateOffer(offerRequest);
 	}
 	
 }